updated application layout
