+++ ui/src/components/buttons/account-popover.js
@@ -1,9 +1,6 @@
 import { useCallback } from 'react';
 import PropTypes from 'prop-types';
 import toast from 'react-hot-toast';
-// import CreditCard01Icon from '@untitled-ui/icons-react/build/esm/CreditCard01';
-// import Settings04Icon from '@untitled-ui/icons-react/build/esm/Settings04';
-// import User03Icon from '@untitled-ui/icons-react/build/esm/User03';
 import Box from '@mui/material/Box';
 import Button from '@mui/material/Button';
 import Divider from '@mui/material/Divider';
@@ -14,7 +11,7 @@ import Popover from '@mui/material/Popover';
 import SvgIcon from '@mui/material/SvgIcon';
 import Typography from '@mui/material/Typography';
 
-import { RouterLink } from '../../components/router-link';
+import { RouterLink } from '../router-link';
 import { useAuth } from '../../hooks/use-auth';
 import { useMockedUser } from '../../hooks/use-mocked-user';
 import { useRouter } from '../../hooks/use-router';
