+++ ui/src/routes/OthersProfile.js
@@ -1,9 +1,9 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
-import MessageChatSquareIcon from './assets/svg/message-chat-square.js';
-import DotsHorizontalIcon from './assets/svg/dots-horizontal.js';
-import Image03Icon from './assets/svg/image-03.js';
-import UserPlus02Icon from './assets/svg/user-plus-02.js';
+import MessageChatSquareIcon from '../assets/svg/message-chat-square.js';
+import DotsHorizontalIcon from '../assets/svg/dots-horizontal.js';
+import Image03Icon from '../assets/svg/image-03.js';
+import UserPlus02Icon from '../assets/svg/user-plus-02.js';
 import Avatar from '@mui/material/Avatar';
 import Box from '@mui/material/Box';
 import Button from '@mui/material/Button';
@@ -17,17 +17,14 @@ import Tabs from '@mui/material/Tabs';
 import Tooltip from '@mui/material/Tooltip';
 import Typography from '@mui/material/Typography';
 import { blueGrey } from '@mui/material/colors';
-// import EditProfileForm from './EditProfileForm';
-import { socialApi } from './social/index.js';
-import { RouterLink } from './components/router-link';
-// import { Seo } from './components/seo';
-import { useMounted } from './hooks/use-mounted';
-// import { usePageView } from './hooks/use-page-view';
-import { Layout as DashboardLayout } from './layouts/layout.js';
-import { paths } from './paths';
-import { SocialConnections } from './social/social-connections';
-import { SocialTimeline } from './social/social-timeline';
-import Projects from './Projects'
+import { socialApi } from '../social/index.js';
+import { RouterLink } from '../components/router-link';
+import { useMounted } from '../hooks/use-mounted';
+import { Layout as DashboardLayout } from '../layouts/layout.js';
+import { paths } from '../paths';
+import { SocialConnections } from '../social/social-connections';
+import { SocialTimeline } from '../social/social-timeline';
+import Projects from './Projects.js'
 // import Modal from '@mui/material/Modal';
 // import Dialog from '@mui/material/Dialog';
 // import DialogActions from '@mui/material/DialogActions';
